+++ test/test.converter.js
@@ -2,8 +2,10 @@ var assert = require('assert');
 var carbone = require('../lib/index');
 var path  = require('path');
 var fs = require('fs');
+var os = require('os');
 var helper = require('../lib/helper');
 var converter = require('../lib/converter');
+var params = require('../lib/params');
 var exec = require('child_process').exec;
 var tempPath = path.join(__dirname, 'temp');
 
@@ -23,6 +25,35 @@ describe('Converter', function () {
   after(function () {
     helper.rmDirRecursive(tempPath);
     carbone.reset();
+  });
+
+  describe('shouldTheFactoryBeRestarted', function () {
+    it('should return not restart LO if factoryMemoryFileSize = 0 or factoryMemoryThreshold = 0', function () {
+      var _params = {
+        factoryMemoryFileSize  : 0,
+        factoryMemoryThreshold : 0
+      };
+      helper.assert(converter.shouldTheFactoryBeRestarted(_params, 1000, 100), false);
+      helper.assert(converter.shouldTheFactoryBeRestarted(_params, 1   , 1000000), false);
+      _params.factoryMemoryFileSize = 10000;
+      helper.assert(converter.shouldTheFactoryBeRestarted(_params, 1   , 100000000), false);
+      _params.factoryMemoryFileSize = 10000;
+      _params.factoryMemoryThreshold = 0;
+      helper.assert(converter.shouldTheFactoryBeRestarted(_params, 1   , 100000000), false);
+      helper.assert(converter.shouldTheFactoryBeRestarted(_params, 100000000, 100000000), false);
+    });
+    it('should restart LO if the threshold is reached', function () {
+      var _params = {
+        factoryMemoryFileSize  : 2,
+        factoryMemoryThreshold : 20
+      };
+      helper.assert(converter.shouldTheFactoryBeRestarted(_params, 1000, 99), false);
+      helper.assert(converter.shouldTheFactoryBeRestarted(_params, 1000, 100), true);
+      helper.assert(converter.shouldTheFactoryBeRestarted(_params, 1000, 101), true);
+      helper.assert(converter.shouldTheFactoryBeRestarted(_params, 10  , 101), true);
+      helper.assert(converter.shouldTheFactoryBeRestarted(_params, 10  , 1  ), true);
+      helper.assert(converter.shouldTheFactoryBeRestarted(_params, 10  , 0  ), false);
+    });
   });
 
   describe('init', function () {
@@ -104,6 +135,23 @@ describe('Converter', function () {
           });
         });
       });
+    });
+    it('should return a link to the output document', function (done) {
+      var _filePath = path.resolve('./test/datasets/test_odt_render_static.odt');
+      var _outputPath = path.join(tempPath, 'output.toto');
+      converter.convertFile(_filePath, 'writer_pdf_Export', '', function (err, result) {
+        helper.assert(err+'', 'null');
+        helper.assert(typeof(result) === 'string', true);
+        helper.assert(_outputPath, result);
+        fs.readFile(_outputPath, function (err, content) {
+          helper.assert(err+'', 'null');
+          assert.equal(content.slice(0, 4).toString(), '%PDF');
+          fs.unlink(_outputPath, function (err) {
+            helper.assert(err+'', 'null');
+            done();
+          });
+        });
+      }, true, _outputPath);
     });
     it('should restart automatically the conversion factory if it crashes', function (done) {
       var _filePath = path.resolve('./test/datasets/test_odt_render_static.odt');
@@ -208,7 +256,7 @@ describe('Converter', function () {
       converter.init({factories : 1, startFactory : true, tempPath : tempPath}, function (factories) {
         var _officePID = factories['0'].pid;
         converter.convertFile(_filePath, 'writer_pdf_Export', '', function (err) {
-          assert.equal(err, 'Could not open document');
+          assert.equal(err+'', 'Error: Could not open document');
           assert.equal(factories['0'].pid, _officePID);
           done();
         });
@@ -219,7 +267,7 @@ describe('Converter', function () {
       converter.init({factories : 1, startFactory : true, tempPath : tempPath}, function (factories) {
         var _officePID = factories['0'].pid;
         converter.convertFile(_filePath, 'MS Word 97', '', function (err) {
-          assert.equal(err, 'Could not convert document');
+          assert.equal(err+'', 'Error: Could not convert document');
           assert.equal(factories['0'].pid, _officePID);
           done();
         });
@@ -283,31 +331,37 @@ describe('Converter', function () {
   });
 
   describe('LO Memory monitoring and report timeout', function () {
+    const totatMemoryAvailableMB     = os.totalmem() / 1024 / 1024;
+    const factoryMemoryThreshold     = 8; // 8 %
+    const nbConversionBoundary       = 5; // below it os ok, above it restarts LO
+    const nbConversionForRestart     = nbConversionBoundary + 1;
+    const nbConversionWithoutRestart = nbConversionBoundary - 1;
+    // adapt factoryMemoryFileSize according to the system memory to make the test machine-independant
+    const factoryMemoryFileSize      = totatMemoryAvailableMB * factoryMemoryThreshold / (nbConversionBoundary * 100);
     const memoryOptions = {
       factories              : 1,
       startFactory           : true,
       tempPath               : tempPath,
-      factoryMemoryCheck     : true,
-      factoryMemoryFileSize  : 20,
-      factoryMemoryThreshold : 5 // tested on a 16GB system
+      factoryMemoryFileSize  : factoryMemoryFileSize,
+      factoryMemoryThreshold : factoryMemoryThreshold
     };
 
     afterEach(function (done) {
       converter.exit(done);
+      carbone.reset();
     });
 
     it('should not reach the memory threshold and the factory process shouldn\'t be killed', function (done) {
       var _filePath = path.resolve('./test/datasets/test_odt_render_static.odt');
       converter.init(memoryOptions, function (factories) {
         const _officePID = factories['0'].pid;
-        const _maxLoops = 10;
-        for (let i = 0; i <= _maxLoops; i++) {
+        for (let i = 0; i <= nbConversionWithoutRestart; i++) {
           converter.convertFile(_filePath, 'writer_pdf_Export', '', function (err, result) {
             helper.assert(err+'', 'null');
             var _buf = new Buffer(result);
             assert.equal(_buf.slice(0, 4).toString(), '%PDF');
             assert.equal(factories['0'].pid, _officePID);
-            if (i === _maxLoops) {
+            if (i === nbConversionWithoutRestart) {
               done();
             }
           });
@@ -319,13 +373,12 @@ describe('Converter', function () {
       var _filePath = path.resolve('./test/datasets/test_odt_render_static.odt');
       converter.init(memoryOptions, function (factories) {
         const _officePID = factories['0'].pid;
-        const _maxLoops = 50;
-        for (let i = 0; i <= _maxLoops; i++) {
+        for (let i = 0; i <= nbConversionForRestart; i++) {
           converter.convertFile(_filePath, 'writer_pdf_Export', '', function (err, result) {
             helper.assert(err+'', 'null');
             var _buf = new Buffer(result);
             assert.equal(_buf.slice(0, 4).toString(), '%PDF');
-            if (i === _maxLoops) {
+            if (i === nbConversionForRestart) {
               assert.notEqual(factories['0'].pid, _officePID);
               done();
             }
@@ -339,7 +392,7 @@ describe('Converter', function () {
       memoryOptions.factories = 4;
       converter.init(memoryOptions, function (factories) {
         const _officePIDs = [factories['0'].pid, factories['1'].pid, factories['2'].pid, factories['3'].pid];
-        const _maxLoops = 200;
+        const _maxLoops = nbConversionForRestart * memoryOptions.factories;
         for (let i = 0; i <= _maxLoops; i++) {
           converter.convertFile(_filePath, 'writer_pdf_Export', '', function (err, result) {
             helper.assert(err+'', 'null');
@@ -371,14 +424,55 @@ describe('Converter', function () {
       });
     });
 
-    it('should timeout and return an error.', function (done) {
+    it('should timeout if the report is too long to convert, return an error and should be able to convert again after', function (done) {
       var _filePath = path.resolve('./test/datasets/test_odt_render_static.odt');
-      memoryOptions.converterFactoryTimeout = 500;
-      converter.init(memoryOptions, function () {
+      memoryOptions.converterFactoryTimeout = 5;
+      converter.init(memoryOptions, function (factories) {
+        const _officePID = factories['0'].pid;
         converter.convertFile(_filePath, 'writer_pdf_Export', '', function (err, result) {
-          helper.assert(err, 'Carbone Converter timeout error: the process is taking too long.');
-          helper.assert(result+'', 'null');
-          done();
+          helper.assert(err instanceof Error, true);
+          helper.assert(err+'', 'Error: Document conversion timeout reached ('+params.converterFactoryTimeout+' ms)');
+          helper.assert(result+'', 'undefined');
+          params.converterFactoryTimeout = 10000;
+          converter.convertFile(_filePath, 'writer_pdf_Export', '', function (err, result) {
+            const _newOfficePIDAfter = factories['0'].pid;
+            helper.assert(_officePID !== _newOfficePIDAfter, true);
+            helper.assert(err+'', 'null');
+            var _buf = new Buffer(result);
+            assert.equal(_buf.slice(0, 4).toString(), '%PDF');
+            done();
+          });
+        });
+      });
+    });
+
+    it('should convert one file, another with timeout, and another without timeout', function (done) {
+      var _filePath = path.resolve('./test/datasets/test_odt_render_static.odt');
+      memoryOptions.converterFactoryTimeout = 5;
+      converter.init(memoryOptions, function (factories) {
+        const _officePID = factories['0'].pid;
+        params.converterFactoryTimeout = 10000;
+        converter.convertFile(_filePath, 'writer_pdf_Export', '', function (err, result) {
+          const _newOfficePID = factories['0'].pid;
+          helper.assert(_officePID, _newOfficePID);
+          helper.assert(err+'', 'null');
+          var _buf = new Buffer(result);
+          assert.equal(_buf.slice(0, 4).toString(), '%PDF');
+          params.converterFactoryTimeout = 2;
+          converter.convertFile(_filePath, 'writer_pdf_Export', '', function (err, result) {
+            helper.assert(err instanceof Error, true);
+            helper.assert(err+'', 'Error: Document conversion timeout reached ('+params.converterFactoryTimeout+' ms)');
+            helper.assert(result+'', 'undefined');
+            params.converterFactoryTimeout = 10000;
+            converter.convertFile(_filePath, 'writer_pdf_Export', '', function (err, result) {
+              const _newOfficePIDAfter = factories['0'].pid;
+              helper.assert(_officePID !== _newOfficePIDAfter, true);
+              helper.assert(err+'', 'null');
+              var _buf = new Buffer(result);
+              assert.equal(_buf.slice(0, 4).toString(), '%PDF');
+              done();
+            });
+          });
         });
       });
     });